Notes-gen-model/requirements.txt
transformers>=4.35.0
datasets>=2.14.0
torch>=2.0.0
tqdm>=4.66.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0
# RAG and Advanced Features
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
peft>=0.6.0
bitsandbytes>=0.41.0
rouge-score>=0.1.2
nltk>=3.8.1
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

